+++ Source/Iterators/Standard/IGeneratorAdapter.php
@@ -6,7 +6,7 @@ use Pinq\Iterators\Generators\IGenerator;
 
 /**
  * Implementation of the adapter iterator for generators using the fetch method.
- * This class providers interopability between the generator scheme and the standard
+ * This class providers interoperability between the generator scheme and the standard
  * scheme.
  *
  * @author Elliot Levin <elliotlevin@hotmail.com>
@@ -16,7 +16,22 @@ class IGeneratorAdapter extends IteratorAdapter
     /**
      * @var IGenerator
      */
-    private $generator;
+    protected $generator;
+
+    /**
+     * @var \Generator
+     */
+    protected $currentGeneratorWrapper;
+
+    /**
+     * @var mixed
+     */
+    protected $currentKey;
+
+    /**
+     * @var mixed
+     */
+    protected $currentValue;
 
     public function __construct(IGenerator $generator)
     {
@@ -26,8 +41,26 @@ class IGeneratorAdapter extends IteratorAdapter
 
     public function doRewind()
     {
-        //IGenerator is an \IteratorAggregate and a new generator must be built for every rewind
-        $this->iterator = $this->generator->getIterator();
+        $this->iterator = $this->buildGeneratorWrapper();
         parent::doRewind();
+    }
+
+    //Considering this is an adapter for generators, they must be supported
+    //hence we should be able to utilise generator in this iterator
+    protected function buildGeneratorWrapper()
+    {
+        $this->preventFirstNext = true;
+        foreach ($this->generator as $key => &$value) {
+            yield;
+            $this->currentKey   = $key;
+            $this->currentValue =& $value;
+        }
+    }
+
+    protected function doFetch()
+    {
+        if (parent::doFetch() !== null) {
+            return [$this->currentKey, &$this->currentValue];
+        }
     }
 }