+++ Source/Providers/Configuration/IQueryConfiguration.php
@@ -6,6 +6,7 @@ use Pinq\Caching;
 use Pinq\Queries\Builders;
 use Pinq\Queries;
 use Pinq\Iterators\IIteratorScheme;
+use Pinq\Providers\Utilities;
 
 /**
  * Interface for the configurable services of the query
@@ -19,6 +20,11 @@ interface IQueryConfiguration
      * @return Caching\IQueryCache
      */
     public function getQueryCache();
+
+    /**
+     * @return Utilities\IQueryResultCollection|null
+     */
+    public function getQueryResultCollection();
 
     /**
      * @return Queries\Builders\IRequestQueryBuilder