+++ Source/Providers/QueryProvider.php
@@ -4,12 +4,11 @@ namespace Pinq\Providers;
 
 use Pinq\Caching;
 use Pinq\Expressions as O;
-use Pinq\Queries\Builders;
 use Pinq\Queries;
+use Pinq\Queries\Builders;
 
 /**
- * Base class for the query provider, with default functionality
- * for the function to expression tree converter and request evaluation
+ * Base class for the query provider.
  *
  * @author Elliot Levin <elliotlevin@hotmail.com>
  */
@@ -20,12 +19,21 @@ abstract class QueryProvider extends ProviderBase implements IQueryProvider
      */
     protected $requestBuilder;
 
-    public function __construct(Queries\ISourceInfo $sourceInfo, Configuration\IQueryConfiguration $configuration = null)
-    {
+    public function __construct(
+            Queries\ISourceInfo $sourceInfo,
+            Configuration\IQueryConfiguration $configuration = null
+    ) {
         parent::__construct($sourceInfo, $configuration ? : new Configuration\DefaultQueryConfiguration());
 
-        $this->requestBuilder = $this->configuration->getRequestQueryBuilder();
+        $this->requestBuilder        = $this->configuration->getRequestQueryBuilder();
+        $this->queryResultCollection = $this->configuration->getQueryResultCollection();
     }
+
+    public function getQueryResultCollection()
+    {
+        return $this->queryResultCollection;
+    }
+
 
     public function createQueryable(O\TraversalExpression $scopeExpression = null)
     {
@@ -33,6 +41,23 @@ abstract class QueryProvider extends ProviderBase implements IQueryProvider
     }
 
     public function load(O\Expression $requestExpression)
+    {
+        if($this->queryResultCollection === null) {
+            return $this->loadRequestExpression($requestExpression);
+        }
+
+        if($this->queryResultCollection->tryComputeResults($requestExpression, $results)) {
+            return $results;
+        }
+
+        $queryExpression = $this->queryResultCollection->optimizeQuery($requestExpression);
+        $results = $this->loadRequestExpression($queryExpression);
+        $this->queryResultCollection->saveResults($queryExpression, $results);
+
+        return $this->queryResultCollection->computeResults($requestExpression);
+    }
+
+    protected function loadRequestExpression(O\Expression $requestExpression)
     {
         $resolution = $this->requestBuilder->resolveRequest($requestExpression);
         $queryHash  = $resolution->getHash();